requirements.txt
openapi==2.0.0
python-dotenv==1.1.0